Msvc warnings (#822)
fix msvc warnings in python bindings and simulations

+++ bindings/python/src/sha1_hash.cpp
@@ -11,7 +11,7 @@
 long get_hash(boost::python::object o)
 {
     using namespace boost::python;
-    return PyObject_Hash(str(o).ptr());
+    return long(PyObject_Hash(str(o).ptr()));
 }
 
 using namespace libtorrent;
